moving config in ccd package
